feat: refactor Footer and Organization components to streamline social media integration and clean up unused props
