+++ shared/react-components/src/sections/Footer/index.tsx
@@ -1,40 +1,23 @@
+import InstagramIcon from '../../icons/Instagram';
+import LinkedinIcon from '../../icons/Linkedin';
 import Paragraph from '../../paragraph';
+import { FooterProps } from './interface';
 
-export interface LinkItem {
-  text: string;
-  url: string;
+const iconsByName: Record<string, typeof InstagramIcon> = {
+  "Instagram": InstagramIcon,
+  "Linkedin": LinkedinIcon
 }
 
-export interface AboutSection {
-  title: string;
-  description: string;
-}
-
-export interface LinksSection {
-  title: string;
-  items: LinkItem[];
-}
-
-export interface ContactSection {
-  title: string;
-  location: string;
-  email: string;
-}
-
-export interface SocialSection {
-  title: string;
-  location: string;
-  email: string;
-}
-
-export interface FooterProps {
-  about: AboutSection;
-  links: LinksSection;
-  contact: ContactSection;
-  social: SocialSection;
-  copyright: string;
-  legalLinks: LinkItem[];
-}
+const socials = [
+  {
+    url: "https://www.instagram.com/devopsdayslima/",
+    iconName: "Instagram",
+  },
+  {
+    url: "https://www.linkedin.com/company/devops-days-lima/",
+    iconName: "Linkedin",
+  }
+]
 
 export default function Footer({
   about,
@@ -78,10 +61,23 @@ export default function Footer({
 
           <div>
             <Paragraph color="tertiary" size="xl" className="mb-4">{social.title}</Paragraph>
-            <Paragraph className="mb-2">{social.location}</Paragraph>
-            <Paragraph as="a" href={`mailto:${social.email}`} className="hover:text-gray-300 transition-colors">
-              {social.email}
-            </Paragraph>
+            <div className='flex gap-6 items-baseline'>
+              {socials.map(({ url, iconName }) => {
+                const Icon = iconsByName[iconName];
+
+                return (
+                  <a
+                    key={url}
+                    href={url}
+                    target="_blank"
+                    rel="noopener noreferrer"
+                    className="text-white"
+                  >
+                    {Icon && <Icon width={32} height={32} />}
+                  </a>
+                );
+              })}
+            </div>
           </div>
         </div>
